requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
groq>=0.4.2
faiss-cpu>=1.7.4
sentence-transformers>=2.5.1
python-dotenv>=1.0.0
Pillow>=10.2.0
requests>=2.31.0
numpy>=1.26.0
